#34: refactoring (renamed file)
