+++ dev/test_trajectory_io.ts
